+++ app/smartbrain_3000/web/_app/immutable/chunks/Cf98RD8Z.js
@@ -1,0 +1,1 @@
+import"./BI6jKVs1.js";